+++ sets/src/index.ts
@@ -1,3 +1,3 @@
-export {PokemonSet} from '@pkmn/types';
+export type {PokemonSet} from '@pkmn/types';
 export {Sets, Data} from './sets';
 export {Team, Teams} from './teams';