+++ sets/src/test/bundle.test.ts
@@ -6,9 +6,9 @@ import {Sets, Teams} from '../index';
 describe('Bundle', () => {
   it('usage', () => {
     {
-      const window = {} as { PokemonSets: typeof Sets; PokemonTeams: typeof Teams };
+      const window = {} as {pkmn: {sets: { Sets: typeof Sets; Teams: typeof Teams }}};
       // eslint-disable-next-line no-eval
-      eval(fs.readFileSync(path.resolve(__dirname, '../../build/production.min.js'), 'utf8'));
+      eval(fs.readFileSync(path.resolve(__dirname, '../../build/index.min.js'), 'utf8'));
       const raw =
 `Tangrowth @ Assault Vest
 Ability: Regenerator
@@ -20,8 +20,8 @@ IVs: 30 Atk / 30 Def
 - Hidden Power [Ice]
 - Earthquake
 `;
-      const set = window.PokemonSets.importSet(raw);
-      const team = window.PokemonTeams.importTeam(raw)!;
+      const set = window.pkmn.sets.Sets.importSet(raw);
+      const team = window.pkmn.sets.Teams.importTeam(raw)!;
       expect(set.ability).toBe('Regenerator');
       expect(set.moves).toContain('Knock Off');
       expect(team.team).toHaveLength(1);