+++ sim/test/main.js
@@ -1,6 +1,6 @@
 'use strict';
 
 before('initialization', function () {
-	global.Dex = require('./../build/sim').Dex;
+	global.Dex = require('./..').Dex;
 	global.toID = Dex.toID;
 });