+++ sim/test/sim/moves/roost.js
@@ -2,7 +2,7 @@
 
 const assert = require('./../../assert');
 const common = require('./../../common');
-const Sim = require('./../../../build/sim');
+const Sim = require('./../../../build/cjs/sim');
 
 let battle;